working on the lakes restart, it also uses marks2017
