+++ awsm/tests/basins/Lakes/config.ini
@@ -60,6 +60,7 @@ wind_ninja_tz:                 UTC
 [precip]
 distribution:                  grid
 grid_local:		               True
+new_snow_density_model:             susong1999
 
 ################################################################################
 # Albedo distribution
