+++ awsm/tests/basins/Lakes/gold_hrrr/gold_config.ini
@@ -104,9 +104,8 @@ storm_days_restart:                 None
 precip_temp_method:                 dew_point
 detrend_slope:                      1
 grid_mask:                          False
-marks2017_timesteps_to_end_storms:  6
 susong1999_timesteps_to_end_storms: 6
-new_snow_density_model:             marks2017
+new_snow_density_model:             susong1999
 precip_rescaling_model:             None
 winstral_veg_3011:                  0.7
 station_undercatch_model_default:   us_nws_8_shielded
